refactor: move HTML data-attribute API to a lib, add more tests
